style: rearranges settings route
