+++ src/app/settings/payments/page.tsx
@@ -92,9 +92,9 @@ export default function Page({}: Props) {
   }
 
   return (
-    <div className="bg-white h-screen flex flex-col overflow-hidden">
+    <div className="bg-white flex flex-col overflow-hidden">
 
-      <div className="p-3 md:px-[50px] w-full space-y-2">
+      <div className="p-3 w-full space-y-2">
         <div className="space-y-6">
         <div className="space-y-6">
           <h1 className="text-lg font-bold">{pageName}</h1>
